feat: split some components to a better readability
